Started change app to Spring boot

+++ src/main/java/ru/milov/transactions/dao/UserBillDao.java
@@ -1,13 +1,13 @@
 package ru.milov.transactions.dao;
 
 import org.springframework.stereotype.Service;
-import ru.milov.transactions.service.domain.UserBill;
+import ru.milov.transactions.service.entity.UserBill;
 import javax.sql.DataSource;
 import java.sql.*;
 import java.util.List;
 
 @Service
-public class UserBillDao implements Dao <UserBill, Integer>{
+public class UserBillDao implements Dao <UserBill, Long>{
 
     private final DataSource dataSource;
 
@@ -21,8 +21,8 @@ public class UserBillDao implements Dao <UserBill, Integer>{
         String request = "select * from bills where user_id = ? and bill_id = ?";
         try (Connection connection = dataSource.getConnection()) {
             PreparedStatement statement = connection.prepareStatement(request);
-            statement.setInt(1, userBill.getUser_id());
-            statement.setInt(1, userBill.getBill_id());
+            statement.setLong(1, userBill.getUser_id());
+            statement.setLong(1, userBill.getBill_id());
 
             ResultSet rSet = statement.executeQuery();
             while (rSet.next()) {
@@ -42,13 +42,13 @@ public class UserBillDao implements Dao <UserBill, Integer>{
 
         try (Connection connection = dataSource.getConnection()) {
             PreparedStatement statement = connection.prepareStatement(request);
-            statement.setInt(1, userBill.getUser_id());
+            statement.setLong(1, userBill.getUser_id());
 
             ResultSet rSet = statement.executeQuery();
             while (rSet.next()) {
                 userBill = new UserBill();
-                userBill.setUser_id(rSet.getInt("user_id"));
-                userBill.setBill_id(rSet.getInt("bill_id"));
+                userBill.setUser_id(rSet.getLong("user_id"));
+                userBill.setBill_id(rSet.getLong("bill_id"));
                 userBill.setName(rSet.getString("name"));
                 userBill.setBalance(rSet.getBigDecimal("balance"));
                 list.add(userBill);
@@ -66,7 +66,7 @@ public class UserBillDao implements Dao <UserBill, Integer>{
         try (Connection connection = dataSource.getConnection()) {
                 PreparedStatement ps = connection.prepareStatement(request,Statement.RETURN_GENERATED_KEYS);
 
-                ps.setInt(1, userBill.getUser_id());
+                ps.setLong(1, userBill.getUser_id());
                 ps.setString(2, userBill.getName());
                 ps.setBigDecimal(3, userBill.getBalance());
 
@@ -79,7 +79,7 @@ public class UserBillDao implements Dao <UserBill, Integer>{
             try (ResultSet generatedKeys = ps.getGeneratedKeys()) {
                 if (generatedKeys.next()) {
 
-                    userBill.setBill_id(generatedKeys.getInt(1));
+                    userBill.setBill_id(generatedKeys.getLong(1));
                 } else {
                     throw new SQLException("Creating bill failed, no ID obtained.");
                 }
@@ -102,8 +102,8 @@ public class UserBillDao implements Dao <UserBill, Integer>{
         try(Connection connection = dataSource.getConnection()) {
             PreparedStatement statement = connection.prepareStatement(request);
             statement.setBigDecimal(1, userBill.getBalance());
-            statement.setInt(2, userBill.getUser_id());
-            statement.setInt(3, userBill.getBill_id());
+            statement.setLong(2, userBill.getUser_id());
+            statement.setLong(3, userBill.getBill_id());
 
             statement.executeUpdate();
         } catch (SQLException throwables) {
@@ -113,12 +113,12 @@ public class UserBillDao implements Dao <UserBill, Integer>{
     }
 
     @Override
-    public boolean delete(Integer bill_id) {
+    public boolean delete(Long bill_id) {
         String request = "delete from bills where bill_id = ?";
         try (Connection connection = dataSource.getConnection()) {
             PreparedStatement ps = connection.prepareStatement(request);
 
-            ps.setInt(1, bill_id);
+            ps.setLong(1, bill_id);
 
             ps.executeUpdate();
